Test cases init

+++ src/models/orders.ts
@@ -20,12 +20,15 @@ export type OrderProduct ={
 
 }
 
+
+//custom type for the custom joint tables query
 export type OrderEntry = {
     product_name: string;
     product_price: number;
     product_quantity: number;
 }
 
+//custom type for a custom joint tables query that gets total price of an order depending on prices and quantities of products in it
 export type OrderPrice = {
     total_price: number;
 }
@@ -35,8 +38,8 @@ export type OrderPrice = {
 
 export class OrderStore
 {
-    //get the total price of an order by order id
-    async getOrderTotalPrice(id: string): Promise<OrderPrice> 
+    //get the total price of an order by order id and user id
+    async getOrderTotalPrice(order_id: string, user_id : string): Promise<OrderPrice> 
     {
 
         try {
@@ -45,8 +48,8 @@ export class OrderStore
             FROM orders INNER JOIN orders_details INNER JOIN products 
             ON  orders_details.product_id = products.id
             ON  orders.id = orders_details.order_id
-            WHERE orders.id = $1`; //sql query
-            const result = await conn.query(sql, [id]); 
+            WHERE orders.id = $1 and orders.user_id = $2`; //sql query
+            const result = await conn.query(sql, [order_id, user_id]); 
             conn.release();
             return result.rows[0];
            } catch (error) {
@@ -55,8 +58,8 @@ export class OrderStore
 
     }
 
-    //get the list of items in an order by order id (whether active or not)
-    async getOrderProducts(id: string): Promise<OrderEntry[]> 
+    //get the list of items(product name, price, quantity) in an order by order_id (whether active or complete) and user id.
+    async getOrderProducts(order_id: string, user_id: string): Promise<OrderEntry[]> 
     {
 
         try {
@@ -65,8 +68,8 @@ export class OrderStore
             FROM orders INNER JOIN orders_details INNER JOIN products 
             ON  orders_details.product_id = products.id
             ON  orders.id = orders_details.order_id
-            WHERE orders.id = $1`; //sql query
-            const result : QueryResult<OrderEntry> = await conn.query(sql, [id]);
+            WHERE orders.id = $1 and orders.user_id = $2`; //sql query
+            const result : QueryResult<OrderEntry> = await conn.query(sql, [order_id, user_id]);
             conn.release();
             return result.rows;
            } catch (error) {
@@ -75,7 +78,7 @@ export class OrderStore
 
     }
 
-    //get all orders (active or complete)
+    //get all orders (without details) and specify what orders to show (active or complete)
     async getOrders(user_id: string, status: string): Promise<Order[]> {
 
         try {
@@ -91,21 +94,29 @@ export class OrderStore
 
 
     //adds a product to as specific order
-    async addProduct(op: OrderProduct): Promise<OrderProduct>{
+    async addProduct(user_id: string , op: OrderProduct): Promise<OrderProduct>{
         try {
             const conn =  await client.connect();
-            const sqlOrder = `SELECT * FROM orders WHERE id = $1`;
-            const resultOrder = await conn.query(sqlOrder, [op.order_id]);
+            const sqlOrder = `SELECT * FROM orders WHERE id = $1 and user_id = $2`;
+            const resultOrder = await conn.query(sqlOrder, [op.order_id, user_id]);
             const order = resultOrder.rows[0];
             if (order.status == 'complete')
             {
                 throw new Error(`can not insert product in a completed order.`);
             }
 
+            conn.release();
+           } catch (error) {
+                throw new Error(`Can't find order ${error}`);
+           }
+
+           try {
+            const conn =  await client.connect();
             const sqlAddProduct = `INSERT INTO orders_details (order_id, product_id, quantity) VALUES ($1, $2, $3) RETURNING *`
             const resultAddProduct = await conn.query(sqlAddProduct, [op.order_id, op.product_id, op.quantity]);
             conn.release();
             return resultAddProduct.rows[0];
+
            } catch (error) {
             throw new Error(`Can't add product to order ${error}`);
            }
@@ -128,12 +139,12 @@ export class OrderStore
     }
 
     //delete order by id
-    async delete(id: string): Promise<Order> 
+    async delete(order_id: string, user_id: string): Promise<Order> 
     { 
         try {
         const conn = await client.connect();
-        const sql = 'DELETE FROM orders WHERE id= $1 RETURNING *';
-        const result = await conn.query(sql, [id])
+        const sql = 'DELETE FROM orders WHERE id= $1 and user_id = $2 RETURNING *';
+        const result = await conn.query(sql, [user_id, order_id])
         conn.release();
         return result.rows[0];
         }
@@ -143,13 +154,13 @@ export class OrderStore
     }
 
     //set status of order, takes id and new status (should be active or complete to match enum type in db)
-    async setStatus(id: string, status: string): Promise<Order>{
+    async setStatus(order_id: string, user_id: string,status: string): Promise<Order>{
 
         try
         {
             const conn = await client.connect();
-            const sql = `UPDATE orders SET status = $1 WHERE ID = $2 RETURNING *`
-            const result = await conn.query(sql, [status , id]);
+            const sql = `UPDATE orders SET status = $1 WHERE id = $2 and user_id = $3 and user RETURNING *`
+            const result = await conn.query(sql, [status , order_id, user_id]);
             conn.release();
             return result.rows[0];
         }
